// src/app/page/master/jenis-pajak/jenis-pajak.interface.ts
    import {EventEmitter} from '@angular/core';

    export interface jenisPajak {
		kdProfile?: string;
		jenisPajak?: string;
		kdJenisPajakHead?: number;
		deskripsi?: string;
		kdDepartemen?: number;
		reportDisplay?: string;
		kodeExternal?: string;
		namaExternal?: string;
		tampilanLaporan?: string;
		noRec?: string;
		kode?:any;
    }